requirements.txt
networkx==3.1
pyvis==0.3.2
python-dotenv==1.0.0
matplotlib==3.7.1
pandas==2.0.3
streamlit==1.25.0
requests==2.31.0
Pillow==9.5.0
numpy==1.24.3
